feat: layouts can now receive partial data accompagnied with a loading flag (``isLoading`` props)
